Switch RPC namespace form ethcore_ to trace_
