+++ rpc/src/v1/tests/mocked/ethcore.rs
@@ -18,18 +18,12 @@ use std::sync::Arc;
 use std::str::FromStr;
 use jsonrpc_core::IoHandler;
 use v1::{Ethcore, EthcoreClient};
-use ethcore::client::{TestBlockChainClient};
 use ethcore::miner::MinerService;
 use v1::tests::helpers::TestMinerService;
 use util::numbers::*;
 use rustc_serialize::hex::FromHex;
 use util::log::RotatingLogger;
 use util::network_settings::NetworkSettings;
-
-fn blockchain_client() -> Arc<TestBlockChainClient> {
-	let client = TestBlockChainClient::new();
-	Arc::new(client)
-}
 
 fn miner_service() -> Arc<TestMinerService> {
 	Arc::new(TestMinerService::default())
@@ -51,15 +45,14 @@ fn settings() -> Arc<NetworkSettings> {
 	})
 }
 
-fn ethcore_client(client: &Arc<TestBlockChainClient>, miner: &Arc<TestMinerService>) -> EthcoreClient<TestBlockChainClient, TestMinerService> {
-	EthcoreClient::new(&client, &miner, logger(), settings())
+fn ethcore_client(miner: &Arc<TestMinerService>) -> EthcoreClient<TestMinerService> {
+	EthcoreClient::new(&miner, logger(), settings())
 }
 
 #[test]
 fn rpc_ethcore_extra_data() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -75,8 +68,7 @@ fn rpc_ethcore_default_extra_data() {
 	use util::ToPretty;
 
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -89,8 +81,7 @@ fn rpc_ethcore_default_extra_data() {
 #[test]
 fn rpc_ethcore_gas_floor_target() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -103,8 +94,7 @@ fn rpc_ethcore_gas_floor_target() {
 #[test]
 fn rpc_ethcore_min_gas_price() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -117,8 +107,7 @@ fn rpc_ethcore_min_gas_price() {
 #[test]
 fn rpc_ethcore_set_min_gas_price() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -132,8 +121,7 @@ fn rpc_ethcore_set_min_gas_price() {
 #[test]
 fn rpc_ethcore_set_gas_floor_target() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -147,8 +135,7 @@ fn rpc_ethcore_set_gas_floor_target() {
 #[test]
 fn rpc_ethcore_set_extra_data() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -162,8 +149,7 @@ fn rpc_ethcore_set_extra_data() {
 #[test]
 fn rpc_ethcore_set_author() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -177,11 +163,10 @@ fn rpc_ethcore_set_author() {
 #[test]
 fn rpc_ethcore_dev_logs() {
 	let miner = miner_service();
-	let client = blockchain_client();
 	let logger = logger();
 	logger.append("a".to_owned());
 	logger.append("b".to_owned());
-	let ethcore = EthcoreClient::new(&client, &miner, logger.clone(), settings()).to_delegate();
+	let ethcore = EthcoreClient::new(&miner, logger.clone(), settings()).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -194,8 +179,7 @@ fn rpc_ethcore_dev_logs() {
 #[test]
 fn rpc_ethcore_dev_logs_levels() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -207,8 +191,7 @@ fn rpc_ethcore_dev_logs_levels() {
 #[test]
 fn rpc_ethcore_set_transactions_limit() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -222,8 +205,7 @@ fn rpc_ethcore_set_transactions_limit() {
 #[test]
 fn rpc_ethcore_transactions_limit() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -236,8 +218,7 @@ fn rpc_ethcore_transactions_limit() {
 #[test]
 fn rpc_ethcore_net_chain() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -250,8 +231,7 @@ fn rpc_ethcore_net_chain() {
 #[test]
 fn rpc_ethcore_net_max_peers() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -264,8 +244,7 @@ fn rpc_ethcore_net_max_peers() {
 #[test]
 fn rpc_ethcore_net_port() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -278,8 +257,7 @@ fn rpc_ethcore_net_port() {
 #[test]
 fn rpc_ethcore_rpc_settings() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
 
@@ -292,8 +270,7 @@ fn rpc_ethcore_rpc_settings() {
 #[test]
 fn rpc_ethcore_node_name() {
 	let miner = miner_service();
-	let client = blockchain_client();
-	let ethcore = ethcore_client(&client, &miner).to_delegate();
+	let ethcore = ethcore_client(&miner).to_delegate();
 	let io = IoHandler::new();
 	io.add_delegate(ethcore);
