+++ rpc/src/v1/traits/ethcore.rs
@@ -72,12 +72,6 @@ pub trait Ethcore: Sized + Send + Sync + 'static {
 	/// Returns default extra data
 	fn default_extra_data(&self, _: Params) -> Result<Value, Error>;
 
-	/// Executes the given call and returns the VM trace for it.
-	fn vm_trace_call(&self, _: Params) -> Result<Value, Error>;
-
-	/// Executes the given call and returns the diff for it.
-	fn state_diff_call(&self, params: Params) -> Result<Value, Error>;
-
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
@@ -99,9 +93,6 @@ pub trait Ethcore: Sized + Send + Sync + 'static {
 		delegate.add_method("ethcore_rpcSettings", Ethcore::rpc_settings);
 		delegate.add_method("ethcore_nodeName", Ethcore::node_name);
 		delegate.add_method("ethcore_defaultExtraData", Ethcore::default_extra_data);
-
-		delegate.add_method("ethcore_vmTraceCall", Ethcore::vm_trace_call);
-		delegate.add_method("ethcore_stateDiffCall", Ethcore::state_diff_call);
 
 		delegate
 	}
