+++ rpc/src/v1/traits/traces.rs
@@ -32,6 +32,12 @@ pub trait Traces: Sized + Send + Sync + 'static {
 	/// Returns all traces produced at given block.
 	fn block_traces(&self, _: Params) -> Result<Value, Error>;
 
+	/// Executes the given call and returns the VM trace for it.
+	fn vm_trace_call(&self, _: Params) -> Result<Value, Error>;
+
+	/// Executes the given call and returns the diff for it.
+	fn state_diff_call(&self, params: Params) -> Result<Value, Error>;
+
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
@@ -39,6 +45,10 @@ pub trait Traces: Sized + Send + Sync + 'static {
 		delegate.add_method("trace_get", Traces::trace);
 		delegate.add_method("trace_transaction", Traces::transaction_traces);
 		delegate.add_method("trace_block", Traces::block_traces);
+
+		delegate.add_method("trace_vmTraceCall", Traces::vm_trace_call);
+		delegate.add_method("trace_stateDiffCall", Traces::state_diff_call);
+
 		delegate
 	}
 }
